backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "apscheduler>=3.10.4",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.1",
    "holidays>=0.56",
    "joblib>=1.4.2",
    "lightgbm>=4.5.0",
    "numpy>=1.26.4",
    "optuna>=3.6.1",
    "pandas>=2.2.2",
    "psycopg[binary]>=3.3.3",
    "pvlib>=0.11.1",
    "scikit-learn>=1.5.2",
    "skforecast>=0.14.0",
    "sqlmodel>=0.0.37",
    "uvicorn>=0.30.6",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]
